pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
    "click>=8.1",
    "httpx>=0.27",
    "html2text>=2024.2",
    "pypdf>=4.0",
    "youtube-transcript-api>=0.6",
    "duckduckgo-search>=6.0",
    "networkx>=3.0",
    "rank-bm25>=0.2.2",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "python-multipart>=0.0.6",
    "pyvis>=0.3.2",
]

[project.scripts]
second-brain = "second_brain.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true
